refactor: Renomear arquivos de componentes para seguir a estrutura de pastas correta
