+++ examples/src/index.jsx
@@ -20,18 +20,29 @@ function App() {
       <IntroBlock />
       <View className="p-4 container max-w-5xl m-auto">
 
-        <Text className="text-2xl font-bold text-slate-900">Actions</Text>
-        <SwapBlock />
 
+        <Text className="card-title text-4xl font-bold text-slate-900">Actions</Text>
+        <SwapBlock />
         <DropdownBlock />
         <ModalBlock />
-        <AlertBlock />
-        <BadgeBlock />
-        <TabBlock />
         <ButtonBlock />
-        <LinkBlock />
+
+        <div className="divider"></div>
+
+        <Text className="card-title text-4xl font-bold text-slate-900">Data Display</Text>
+        <BadgeBlock />
         <CarouselBlock />
+
+        <div className="divider"></div>
+
+        <Text className="text-4xl font-bold text-slate-900">Feedback</Text>
+        <AlertBlock />
         <ProgressBlock />
+
+        <div className="divider"></div>
+
+        <TabBlock />
+        <LinkBlock />
       </View>
     </Page>
   );
