+++ src/components/Actions/Modal/modal.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import CommonProps from '../commonProps';
+import CommonProps from '../../commonProps';
 
 interface ModalProps extends CommonProps {
   backdrop?:boolean
