+++ src/components/DataDisplay/Avatar/avatar.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import CommonProps from '../commonProps'
+import CommonProps from '../../commonProps'
 
 interface AvatarProps extends CommonProps {}
